requirements.txt
openai==1.58.1
pandas==2.2.3
python-dotenv==1.0.1
streamlit==1.41.1
